+++ DB/dbo/Tables/Sync Tables/_default_constraints_history.sql
@@ -1,8 +1,8 @@
 CREATE TABLE dbo._default_constraints_history (
     _DatabaseID         int           NOT NULL,
-    _ObjectID           int           NOT NULL,
-    _ParentObjectID     int           NOT NULL,
-    _ParentColumnID     int           NOT NULL,
+    _ObjectID           bigint        NOT NULL,
+    _ParentObjectID     bigint        NOT NULL,
+    _ParentColumnID     bigint        NOT NULL,
     --
     _InsertDate         datetime2     NOT NULL,
     _ModifyDate         datetime2     NOT NULL,
